Create of given number(uding foor loop)
A random number table is a series of digits (0 to 9) arranged randomly in rows and columns, as demonstrated in the small sample shown below. The table usually contains 5-digit numbers, arranged in rows and columns, for ease of reading. Typically, a full table may extend over as many as four or more page