Create Fqactorial of given number
A factorial is product of all the numbers from 1 to n, where n is the user specified number. ... A factorial is a product of all the numbers from 1 to n, where n is the user specified number. If condition statement is used to check the value of 'num' variable is less than or equal to 0.